Remove old fileManager and rename fileManager2
