+++ src/renderer/services/fileManager/FileErrors.ts
